[FIXED JENKINS-29527] - AuthorizationMatrixProperty converter: Inheritance blocking may be passed incorrectly
If "useProjectSecurity" appears in the marshalled stream before "blocksInheritance", the second option may be ignored => the project may start taking inherited permissions. However, seems the code just iterates through fields => we should not use the else clause. The assigned "prop" variable is never used, so it's another clue of the bug