Allow checkName for property descriptor to be invoked from Items.
In the event that an Item's view is acting as a proxy for the
configuration of a project's AuthorizationMatrixProperty, it should
be able to check the name on that view.

If the check fails, submitted configuration includes an extra
permission field ("q": "") which causes a class cast error because
the value is expected to be true or false.