[JENKINS-29815] Add disabling of permission inheritance for Folders
This adds the same tick-box to disable inheritance of global permissions
to Folders as already exists for Projects. Consistency!

Support for the Folder plugin seems to originally have been done by
copy-pasting in an only slightly modified version of an ancient copy of
the AuthorizationMatrixProperty.java file. This commit is as such very
similar to the patch Kohsuke once applied to the Project's AMP:
6411fffb41e4e7b217d0e841462ef23a8e62c45a

Now, these two files are close to identical, and should be refactored
into one class in my opinion:

    src/main/java/hudson/security/AuthorizationMatrixProperty.java
    src/main/java/com/cloudbees/hudson/plugins/folder/properties/AuthorizationMatrixProperty.java