Prepare for Java 25 release Sep 16, 2025
Java 25 is the next Java Long Term Support release.  It is scheduled to
release Sep 16, 2025.  The Jenkins project intends to support new Java
LTS releases soon after they are available.  The initial goal is to have
Jenkins core and the top 200 plugins compiling and testing with Java 25
by Oct 31, 2025.

Without these changes, the plugin does not compile with Java 24.  Since we
can expect that Java 25 will be a superset of Java 24, this update will
help prepare the plugin to eventually be compiled and tested with Java 25.

The changes include:

* Update parent pom from 5.5 to 5.18 (most recent parent pom)
* Remove 3 spotbugs suppressions that are no longer necessary

Testing done:

Confirmed that `mvn clean verify` passes with the changes.